refactor remove selection-api from slobs 0.16.3
