+++ app/services/selection/index.ts
@@ -1,2 +1,14 @@
 export * from './selection';
-export * from './selection-api';
+import {
+  ISceneItemNode,
+} from 'services/scenes';
+
+export interface ISelectionState {
+  selectedIds: string[];
+  lastSelectedId: string;
+}
+
+/**
+ * list of ISceneNode.id or ISceneNode
+ */
+export type TNodesList = string | string[] | ISceneItemNode | ISceneItemNode[];
