+++ test/api/clipboard.ts
@@ -3,7 +3,7 @@ import { useSpectron } from '../helpers/spectron';
 import { getClient } from '../helpers/api-client';
 import { SceneBuilder } from '../helpers/scene-builder';
 import { ISceneItemApi, ISceneNodeApi, IScenesServiceApi } from 'services/scenes';
-import { ISelectionServiceApi } from 'services/selection';
+import { SelectionService } from 'services/selection';
 import { IClipboardServiceApi } from 'services/clipboard';
 import { ISceneCollectionsServiceApi } from 'services/scene-collections';
 import { ISourcesServiceApi } from 'services/sources';
@@ -14,7 +14,7 @@ useSpectron({ restartAppAfterEachTest: false, afterStartCb: afterStart });
 let sceneBuilder: SceneBuilder;
 let getNode: (name: string) => ISceneNodeApi;
 let getNodeId: (name: string) => string;
-let selectionService: ISelectionServiceApi;
+let selectionService: SelectionService;
 let clipboardService: IClipboardServiceApi;
 let sourceFiltersService: SourceFiltersService;
 let sceneCollectionsService: ISceneCollectionsServiceApi;
